export stringify, some custom driver may rely on it, for example the http driver server
